move testing-types out of integration folder
